+++ SysBot.Pokemon/Helpers/DudeBot.cs
@@ -3,7 +3,7 @@ namespace SysBot.Pokemon.Helpers
     public static class DudeBot
     {
         public static string Name { get; set; } = "DudeBot";
-        public const string Version = "v6.5.12";
+        public const string Version = "v6.5.13";
         public const string ConfigPath = "config.json";
     }
 }