+++ SysBot.Pokemon/Helpers/DudeBot.cs
@@ -3,7 +3,7 @@ namespace SysBot.Pokemon.Helpers
     public static class DudeBot
     {
         public static string Name { get; set; } = "DudeBot";
-        public static string Version => "6.5.14";
+        public static string Version => "6.5.15";
         public const string ConfigPath = "config.json";
     }
 }